Hoàn thành đến học vấn

+++ backend/app/Http/Controllers/SkillController.php
@@ -8,54 +8,63 @@ use App\Models\Skill;
 class SkillController extends Controller
 {
     // Thêm thông tin skill 
-    public function store(Request $request)
+    public function storeOrUpdate(Request $request)
     {
-        $validated = $request->validate([
-            'skill_info' => 'string|nullable',
-        ]);
-        // Tạo thông tin
-        $skill = Skill::create($validated);
-
-        return response()->json([
-            'success' => true,
-            'message' => 'Thêm thông tin kỹ năng thành công!',
-            'data' => $skill,
-        ]);
-    }
-
-    // Cập nhật skill 
-    public function update(Request $request, $id)
-    {
-        $skill = $request->validate([
-            'skill_info' => 'nullable|string',
+        $request->validate([
+            'skill_info' => 'nullable|string|nullable',
         ]);
 
-        // Kiểm tra bản ghi 
-        $skill = Skill::findOrFail($id);
+        try {
+            $username = $request->user()->username;
 
-        // Cập nhật 
-        $skill->update([
-            'skill_info' => $request->skill_info,
-        ]);
+            $skill = Skill::where("username", $username)->first();
+            $dataToSave = ['skill_info' => $request->skill_info];
 
-        return response()->json([
-            'success' => true,
-            'message' => 'Cập nhật thông tin kỹ năng thành công!',
-            'data' => $skill,
-        ]);
+            $result = Skill::updateOrCreate(
+                ['username' => $username],
+                $dataToSave,
+            );
+
+            return response()->json([
+                'success' => true,
+                'message' => $skill ? 'Cập nhật thông tin kỹ năng thành công!' : "Thêm thông tin kỹ năng thành công!",
+                'data' => $result,
+            ], 200);
+        } catch (\Throwable $th) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Lỗi khi thêm hoặc cập nhật dữ liệu!' . $th->getMessage(),
+            ]);
+        }
     }
 
     // Lấy thông tin skill
-    public function index()
+    public function index(Request $request, $username)
     {
-        return Skill::latest()->get();
+        // Tìm bản ghi cần lấy 
+        $skill = Skill::where('username', $username)->first();
+
+        if (!$skill) {
+            return response()->json([
+                'message' => "Không timg thấy thông tin nào!",
+            ], 404);
+        }
+
+        return $skill;
     }
 
     // Xóa thông tin kỹ năng
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         // Tìm bản ghi cần xóa
         $skill = Skill::findOrFail($id);
+
+        if ($skill->username !== $request->user()->username) {
+            return response()->json([
+                'success' => false,
+                'message' => "Bạn khôn có quyền xóa thông tin này!",
+            ]);
+        }
 
         // Xóa bản ghi trong DB 
         $skill->delete();
